separating out graph functions from final reducer
